prepare-iaas: remove hosts from ssh trust database, and turn off StrictHostKeyChecking
These scripts are run on newly installed machines, and the only time
that there is an entry in the ssh trust database allready is if you are
reinstalling a host with the same name. Therefor it makes sense to
remove the entry from the trustdb before proceeding.

Since we want the script to non-interactive by default,
StrictHostKeyChecking is set to off by default. If you want to manually
verify the host keys before putting them in the trust database, you can
run the script with the -s|--strict option.

This patch also improves the help message, and adds a -h|--help option.