New path for spawned proceses
In Debian 13 (and Ubuntu 24.04) processes are spawned from a systemd in a
diffrent path (/usr/lib/systemd/systemd).

Both the new and old path exist (hard links) on older releases so I think the
easiest and most portable fix is to be less strict when checking and killing
the process.